feat: remove target module and integrate ProofTargets into proof module
